fixing bug on check for target method
during refactor of DrawBattlefield I've found a bug: if a Character is on grid's border the CheckCloseTargets(Grid battlefield) method was throwing an exeption when tring to access a position outside the grid.
Since I've born in august my task would be to: The character can attack/walk in 8 directions. to do that one method that I would implement is to get a valid neibouthood, and this method would help to solve this bug.
An other bug that I've found is that on CheckCloseTarget() method Up and Down bool were getting each other's value